feat(HTD-51): add useDrawerState
